popover on plugin update & a bit of cleanup
