+++ views/popovers/upgrade-tasks.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Monthly badges popover.
+ *
+ * @package Progress_Planner
+ */
+
+use Progress_Planner\Badges\Monthly;
+
+// Exit if accessed directly.
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+?>
+
+<?php
+\progress_planner()->the_view( 'popovers/parts/upgrade-tasks.php', [ 'context' => 'upgrade' ] );
